requirements.txt
typer>=0.9.0
rich>=13.0.0
requests>=2.31.0
google-genai>=0.2.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
